+++ Assets/Scripts/Script/CardEffectFactory/KeyWordEffects/Engage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05ce0f148ee273544bd40e5ea5a7386a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
